Create resource package
